Separate monolithic CLI into different components
