+++ cli/__init__.py
@@ -1,0 +1,2 @@
+"""BinaryAlert release version"""
+VERSION = '1.1.0'
