--- tests/manage_test.py
@@ -1,386 +1,0 @@
-"""Unit tests for the manage.py CLI script."""
-# pylint: disable=no-self-use,protected-access
-import base64
-import getpass
-import inspect
-import os
-import subprocess
-import sys
-from unittest import mock, TestCase
-
-import boto3
-from pyfakefs import fake_filesystem_unittest
-
-import manage
-
-
-def _mock_input(prompt: str) -> str:
-    """Mock for the user input() function to automatically respond with valid answers."""
-    # pylint: disable=too-many-return-statements
-    if prompt.startswith('AWS Account'):
-        return '111122223333'
-    if prompt.startswith('AWS Region'):
-        return 'us-west-2'
-    if prompt.startswith('Unique name prefix'):
-        return ' NEW_NAME_PREFIX '  # Spaces and case shouldn't matter.
-    if prompt.startswith('Enable the CarbonBlack downloader'):
-        return 'yes'
-    if prompt.startswith('CarbonBlack URL'):
-        return 'https://new-example.com'
-    if prompt.startswith('Change the CarbonBlack API token'):
-        return 'yes'
-    if prompt.startswith('Delete all S3 objects'):
-        return 'yes'
-    return 'UNKNOWN'
-
-
-class FakeFilesystemBase(fake_filesystem_unittest.TestCase):
-    """Base class sets up a fake filesystem for other test classes."""
-
-    @staticmethod
-    def _write_config(
-            account_id: str = '123412341234',
-            region: str = 'us-test-1',
-            prefix: str = 'test_prefix',
-            enable_downloader: bool = True,
-            cb_url: str = 'https://cb-example.com',
-            encrypted_api_token: str = 'A'*100):
-        """Create terraform.tfvars file with the given configuration values."""
-        with open(manage.CONFIG_FILE, 'w') as config_file:
-            config_file.write('\n'.join([
-                '// comment1',
-                'aws_account_id = "{}"'.format(account_id),
-                'aws_region = "{}" // comment2'.format(region),
-                'name_prefix = "{}" // comment3'.format(prefix),
-                'enable_carbon_black_downloader = {}'.format(1 if enable_downloader else 0),
-                'carbon_black_url = "{}" //comment4'.format(cb_url),
-                'encrypted_carbon_black_api_token = "{}"'.format(encrypted_api_token),
-                'force_destroy = false',
-                '// comment5'
-            ]))
-
-    def setUp(self):
-        """Enable pyfakefs and write out Terraform config files."""
-        # pylint: disable=no-member
-        self.setUpPyfakefs()
-
-        # pyhcl automatically writes "parsetab.dat" in its site-package path.
-        for path in sys.path:
-            if path.endswith('site-packages'):
-                self.fs.MakeDirectories(os.path.join(path, 'hcl'))
-
-        # Create variables.tf file (and terraform/ directory).
-        self.fs.CreateFile(
-            manage.VARIABLES_FILE,
-            contents='\n'.join([
-                'variable "aws_account_id" {}',
-                'variable "aws_region" {}',
-                'variable "name_prefix" {}',
-                'variable "enable_carbon_black_downloader" {}',
-                'variable "carbon_black_url" {}',
-                'variable "encrypted_carbon_black_api_token" {}'
-            ])
-        )
-
-        # Create terraform.tfvars file.
-        self._write_config()
-
-
-@mock.patch.object(sys, 'stderr', mock.MagicMock())  # pyhcl complains about unused tokens
-class BinaryAlertConfigTestFakeFilesystem(FakeFilesystemBase):
-    """Tests of the BinaryAlertConfig class that use a fake filesystem."""
-
-    def test_property_accesses(self):
-        """Access each property in the BinaryAlertConfig."""
-        config = manage.BinaryAlertConfig()
-
-        self.assertEqual('123412341234', config.aws_account_id)
-        self.assertEqual('us-test-1', config.aws_region)
-        self.assertEqual('test_prefix', config.name_prefix)
-        self.assertEqual(1, config.enable_carbon_black_downloader)
-        self.assertEqual('https://cb-example.com', config.carbon_black_url)
-        self.assertEqual('A' * 100, config.encrypted_carbon_black_api_token)
-        self.assertEqual('test_prefix_binaryalert_batcher', config.binaryalert_batcher_name)
-        self.assertEqual('test.prefix.binaryalert-binaries.us-test-1',
-                         config.binaryalert_s3_bucket_name)
-
-    def test_variable_not_defined(self):
-        """InvalidConfigError is raised if a variable declaration is missing."""
-        with open(manage.CONFIG_FILE, 'w') as config_file:
-            config_file.write('aws_region = "us-east-1"\n')
-
-        with self.assertRaises(manage.InvalidConfigError):
-            manage.BinaryAlertConfig()
-
-    def test_invalid_aws_account_id(self):
-        """InvalidConfigError raised if AWS account ID is not a 12-digit number"""
-        config = manage.BinaryAlertConfig()
-        with self.assertRaises(manage.InvalidConfigError):
-            config.aws_account_id = '1234'
-
-    def test_invalid_aws_region(self):
-        """InvalidConfigError raised if AWS region is set incorrectly."""
-        config = manage.BinaryAlertConfig()
-        with self.assertRaises(manage.InvalidConfigError):
-            config.aws_region = 'us-east-1-'
-
-    def test_invalid_name_prefix(self):
-        """InvalidConfigError raised if name prefix is blank."""
-        config = manage.BinaryAlertConfig()
-        with self.assertRaises(manage.InvalidConfigError):
-            config.name_prefix = ""
-
-    def test_invalid_enable_carbon_black_downloader(self):
-        """InvalidConfigError raised if enable_downloader is not an int."""
-        config = manage.BinaryAlertConfig()
-        with self.assertRaises(manage.InvalidConfigError):
-            config.enable_carbon_black_downloader = '1'
-
-    def test_invalid_carbon_black_url(self):
-        """InvalidConfigError raised if URL doesn't start with http(s)."""
-        config = manage.BinaryAlertConfig()
-        with self.assertRaises(manage.InvalidConfigError):
-            config.carbon_black_url = 'example.com'
-
-    def test_invalid_encrypted_carbon_black_api_token(self):
-        """InvalidConfigError raised if encrypted token is too short."""
-        config = manage.BinaryAlertConfig()
-        with self.assertRaises(manage.InvalidConfigError):
-            config.encrypted_carbon_black_api_token = 'ABCD'
-
-    @mock.patch.object(manage, 'input', side_effect=_mock_input)
-    @mock.patch.object(manage.BinaryAlertConfig, '_encrypt_cb_api_token')
-    def test_configure_with_defaults(
-            self, mock_encrypt: mock.MagicMock, mock_input: mock.MagicMock):
-        """Test configure() when all variables have already had set values."""
-        config = manage.BinaryAlertConfig()
-        config.configure()
-
-        # Verify the mock calls.
-        mock_encrypt.assert_called_once()
-        mock_input.assert_has_calls([
-            mock.call('AWS Region (us-test-1): '),
-            mock.call('Unique name prefix, e.g. "company_team" (test_prefix): '),
-            mock.call('Enable the CarbonBlack downloader? (yes): '),
-            mock.call('CarbonBlack URL (https://cb-example.com): '),
-            mock.call('Change the CarbonBlack API token? (no): ')
-        ])
-
-        # Verify that the configuration has changed.
-        self.assertEqual('us-west-2', config.aws_region)
-        self.assertEqual('new_name_prefix', config.name_prefix)
-        self.assertEqual(1, config.enable_carbon_black_downloader)
-
-    @mock.patch.object(manage, 'input', side_effect=_mock_input)
-    @mock.patch.object(manage.BinaryAlertConfig, '_encrypt_cb_api_token')
-    def test_configure_with_no_defaults(
-            self, mock_encrypt: mock.MagicMock, mock_input: mock.MagicMock):
-        """Test configure() without any values set - no defaults should print."""
-        self._write_config(
-            region='', prefix='', enable_downloader=False, cb_url='', encrypted_api_token=''
-        )
-        config = manage.BinaryAlertConfig()
-        config.configure()
-
-        # Verify the mock calls.
-        mock_encrypt.assert_called_once()
-        mock_input.assert_has_calls([
-            mock.call('AWS Region: '),
-            mock.call('Unique name prefix, e.g. "company_team": '),
-            mock.call('Enable the CarbonBlack downloader? (no): '),
-            mock.call('CarbonBlack URL: '),
-        ])
-
-    def test_validate_valid_with_downloader(self):
-        """Test validate() with all values set correctly."""
-        config = manage.BinaryAlertConfig()
-        config.validate()
-
-        # None of the instance properties should have changed.
-        self.test_property_accesses()
-
-    def test_validate_valid_without_downloader(self):
-        """Test validate() without any CarbonBlack values set - still valid."""
-        self._write_config(enable_downloader=False, cb_url='', encrypted_api_token='')
-        config = manage.BinaryAlertConfig()
-        config.validate()
-
-    def test_validate_invalid(self):
-        """Test validate() with an invalid configuration file."""
-        self._write_config(region='BAD_REGION')
-        config = manage.BinaryAlertConfig()
-
-        with self.assertRaises(manage.InvalidConfigError):
-            config.validate()
-
-    def test_save(self):
-        """New configuration is successfully written and comments are preserved."""
-        config = manage.BinaryAlertConfig()
-        config._config['force_destroy'] = True
-        config.aws_region = 'us-west-2'
-        config.name_prefix = 'new_name_prefix'
-        config.enable_carbon_black_downloader = 0
-        config.carbon_black_url = 'https://example2.com'
-        config.encrypted_carbon_black_api_token = 'B' * 100
-        config.save()
-
-        # Verify that all of the original comments were preserved.
-        with open(manage.CONFIG_FILE) as config_file:
-            raw_data = config_file.read()
-            for i in range(1, 6):
-                self.assertIn('comment{}'.format(i), raw_data)
-
-        new_config = manage.BinaryAlertConfig()
-        self.assertEqual(True, new_config._config['force_destroy'])
-        self.assertEqual(config.aws_region, new_config.aws_region)
-        self.assertEqual(config.name_prefix, new_config.name_prefix)
-        self.assertEqual(
-            config.enable_carbon_black_downloader, new_config.enable_carbon_black_downloader)
-        self.assertEqual(
-            config.encrypted_carbon_black_api_token, new_config.encrypted_carbon_black_api_token)
-
-
-class BinaryAlertConfigTestRealFilesystem(TestCase):
-    """Tests of the BinaryAlertConfig class that use a real filesystem."""
-
-    @mock.patch.object(boto3, 'client')
-    @mock.patch.object(getpass, 'getpass', return_value='abcd' * 10)
-    @mock.patch.object(manage, 'print')
-    @mock.patch.object(subprocess, 'check_call')
-    def test_encrypt_cb_api_token(
-            self, mock_subprocess: mock.MagicMock, mock_print: mock.MagicMock,
-            mock_getpass: mock.MagicMock, mock_client: mock.MagicMock):
-        """Verify that token encryption is done correctly."""
-        mock_client('kms').encrypt.return_value = {'CiphertextBlob': base64.b64encode(b'a'*50)}
-        config = manage.BinaryAlertConfig()
-        config._encrypt_cb_api_token()
-
-        # Verify that the mocks were called as expected.
-        mock_client.assert_has_calls([
-            mock.call().encrypt(KeyId=mock.ANY, Plaintext=mock_getpass.return_value)
-        ])
-        mock_getpass.assert_called_once()
-        mock_print.assert_has_calls([
-            mock.call('Terraforming KMS key...'),
-            mock.call('Encrypting API token...')
-        ])
-        mock_subprocess.assert_has_calls([
-            mock.call(['terraform', 'init']),
-            mock.call(['terraform', 'apply', '-target={}'.format(manage.CB_KMS_ALIAS_TERRAFORM_ID)])
-        ])
-
-
-class ManagerTest(FakeFilesystemBase):
-    """Tests for the Manager class."""
-
-    @mock.patch('sys.stderr', mock.MagicMock())  # pyhcl complains about unused tokens to stderr.
-    def setUp(self):
-        super().setUp()
-        self.manager = manage.Manager()
-
-    def test_commands(self):
-        """Each command should be a function in the class."""
-        for command in self.manager.commands:
-            self.assertTrue(hasattr(self.manager, command))
-
-    def test_help(self):
-        """Help string should contain as many lines as there are commands."""
-        self.assertEqual(len(self.manager.commands), len(self.manager.help.split('\n')))
-
-    @mock.patch.object(boto3, 'client')
-    @mock.patch.object(manage, 'print')
-    def test_analyze_all(self, mock_print: mock.MagicMock, mock_client: mock.MagicMock):
-        """Batch analysis invocation."""
-        self.manager.analyze_all()
-        mock_print.assert_has_calls([
-            mock.call('Asynchronously invoking test_prefix_binaryalert_batcher...'),
-            mock.call('Batcher invocation successful!')
-        ])
-        mock_client.assert_has_calls([
-            mock.call('lambda').invoke(
-                FunctionName='test_prefix_binaryalert_batcher',
-                InvocationType='Event',
-                Qualifier='Production'
-            )
-        ])
-
-    @mock.patch.object(subprocess, 'check_call')
-    def test_apply(self, mock_subprocess: mock.MagicMock):
-        """Validate order of Terraform operations."""
-        self.manager.apply()
-        mock_subprocess.assert_has_calls([
-            mock.call(['terraform', 'init']),
-            mock.call(['terraform', 'fmt']),
-            mock.call(['terraform', 'apply', '-auto-approve=false'])
-        ])
-
-    @mock.patch.object(manage, 'lambda_build')
-    def test_build(self, mock_build: mock.MagicMock):
-        """Calls lambda_build function (tested elsewhere)."""
-        self.manager.build()
-        mock_build.assert_called_once()
-
-    def test_cb_copy_all_not_enabled(self):
-        """Raises InvalidConfigError if the downloader is not enabled."""
-        self._write_config(enable_downloader=False)
-        self.manager = manage.Manager()  # Reload manager with the new config.
-        with self.assertRaises(manage.InvalidConfigError):
-            self.manager.cb_copy_all()
-
-    @mock.patch.object(manage.clone_rules, 'clone_remote_rules')
-    def test_clone_rules(self, mock_clone: mock.MagicMock):
-        """Calls clone_remote_rules (tested elsewhere)."""
-        self.manager.clone_rules()
-        mock_clone.assert_called_once()
-
-    @mock.patch.object(manage.compile_rules, 'compile_rules')
-    @mock.patch.object(manage, 'print')
-    def test_compile_rules(self, mock_print: mock.MagicMock, mock_compile: mock.MagicMock):
-        """Calls compile_rules (tested elsewhere)."""
-        self.manager.compile_rules()
-        mock_compile.assert_called_once()
-        mock_print.assert_called_once()
-
-    @mock.patch.object(manage.BinaryAlertConfig, 'configure')
-    @mock.patch.object(manage, 'print')
-    def test_configure(self, mock_print: mock.MagicMock, mock_configure: mock.MagicMock):
-        """Calls BinaryAlertConfig:configure() (tested elsewhere)."""
-        self.manager.configure()
-        mock_configure.assert_called_once()
-        mock_print.assert_called_once()
-
-    @mock.patch.object(manage.Manager, 'unit_test')
-    @mock.patch.object(manage.Manager, 'build')
-    @mock.patch.object(manage.Manager, 'apply')
-    @mock.patch.object(manage.Manager, 'analyze_all')
-    def test_deploy(self, mock_analyze: mock.MagicMock, mock_apply: mock.MagicMock,
-                    mock_build: mock.MagicMock, mock_test: mock.MagicMock):
-        """Deploy docstring includes each executed command and runs each."""
-        for command in ['unit_test', 'build', 'apply', 'analyze_all']:
-            self.assertIn(command, inspect.getdoc(manage.Manager.deploy))
-
-        self.manager.deploy()
-        mock_test.assert_called_once()
-        mock_build.assert_called_once()
-        mock_apply.assert_called_once()
-        mock_analyze.assert_called_once()
-
-    @mock.patch.object(manage, 'input', side_effect=_mock_input)
-    @mock.patch.object(manage, 'print')
-    @mock.patch.object(subprocess, 'call')
-    @mock.patch.object(subprocess, 'check_call')
-    def test_destroy(self, mock_check_call: mock.MagicMock, mock_call: mock.MagicMock,
-                     mock_print: mock.MagicMock, mock_input: mock.MagicMock):
-        """Destroy asks whether S3 objects should also be deleted."""
-        self.manager.destroy()
-        mock_input.assert_called_once()
-        mock_print.assert_called_once()
-        mock_check_call.assert_called_once()
-        mock_call.assert_called_once()
-
-    @mock.patch.object(manage.live_test, 'run', return_value=False)
-    def test_live_test(self, mock_live_test: mock.MagicMock):
-        """Live test wrapper raises TestFailureError if appropriate."""
-        with self.assertRaises(manage.TestFailureError):
-            self.manager.live_test()
-        mock_live_test.assert_called_once()
